refactor: Renamed incorrest names of folders and solved conflicts with imports
