+++ src/app/shared/interfaces/query-file.interface.ts
