+++ src/app/shared/services/user-info.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { User } from '../interface/user';
+import { User } from '../interfaces/user';
 
 @Injectable({
   providedIn: 'root'
